Adding bidding functionality
- New ItemUser model to allow creating new pivot recodrs
- Item-User ManyToMany relationship
- Policy 'bid' method to allow bidding only once and to disable bidding for your own item
- Minor changes for migrations to have integer values for 'price' fields
- Displaying bid form to input your price

+++ database/migrations/2022_02_18_131440_create_items_table.php
@@ -30,8 +30,8 @@ class CreateItemsTable extends Migration
 
             $table->string('name');
             $table->text('description');
-            $table->decimal('starting_price', 10, 2);
-            $table->decimal('final_price', 10, 2)->nullable();
+            $table->integer('starting_price');
+            $table->integer('final_price')->nullable();
             $table->string('payment_method')->nullable();
             $table->string('delivery_method');
             $table->string('status')->default('active');
